requirements.txt
astroid==2.3.3
autopep8==1.4.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pycryptodome==3.9.4
pylint==2.4.4
rope==0.14.0
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
